requirements.txt
python-telegram-bot>=20.0
playwright
APScheduler
python-dotenv